[BUGFIX] Avoid duplicate calls to update reference index
This patch adds a check when processing the update
reference index stack, in order to avoid calling the
update function more than once with the same input
arguments (which are $table and $uid).

Saves around 100 SQL queries when doing a standard
copy of 25 (nested) pages with 25 content elements.

Change-Id: If162218d4eabb749a92270e18c7c39108fd15f34
Resolves: #78761
Releases: master, 7.6
Reviewed-on: https://review.typo3.org/50724
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>