[SECURITY] Disallow invalid encoding in GeneralUtility::validPathStr
Directory names, which have an invalid UTF encoding,
cause the preg_match() to return false.
To avoid that the complete statement in GeneralUtility::validPathStr()
returns true in this case, a strict comparison against 0 is added,
so that we ensure that strings with invalid encodings are rejected
by this API method.

As a consequence UTF-16 encoded path names are rejected as well, if the
system / file system does not support them.

Resolves: #73453
Releases: master, 8.4, 7.6, 6.2
Security-Commit: 4a1fcf1578ec7d120dac3d2ef68c0267157fd2b4
Security-Bulletins: TYPO3-CORE-SA-2016-023, 024
Change-Id: I4cd6afb4dd2dfe5334dd497271914fb943adfd79
Reviewed-on: https://review.typo3.org/50738
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>