[TASK] Remove compatiblity code added for security release
Remove the overhead that was added to avoid BC breaks for
extension code that subclassed the form view helper.

Enough time is now passed so that extensions can adopt their
subclasses so that the security related hidden field is added as well.

These adaptions will then be compatible with all TYPO3 versions
so that this change here is not marked as breaking.

Resolves: #78869
Releases: master
Change-Id: I910bc26cd57b7629e57332fdab3d57032f0c2478
Reviewed-on: https://review.typo3.org/50863
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>