[BUGFIX] Skip swapping/publishing of deleted records
In case a content element and the accordant page have been deleted in
separate actions and get published together, the workspace process will
trigger an error message since the removed content element cannot be
published anymore (since it has been processed already with the page).

To avoid this behavior deleted records are collected and checked in the
workspace swapping/publishing process.

Change-Id: If04a198abf81efdc88e75da79da0c01cfaa361ff
Resolves: #47384
Releases: master, 7.6
Reviewed-on: https://review.typo3.org/50877
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>