[BUGFIX] Reset hidden field information in FormViewhelper
If a checkbox or a multiple selectbox is rendered, a hidden field is
automatically inserted before to send an empty default value. Field
information are stored within an own viewHelperVariableContainer
variable. After a form is rendered, this information has to be reset to
be able to render another form with same field names. This patch adds
the missing reset of the hidden field information.

Resolves: #78551
Releases: master, 7.6, 6.2
Change-Id: I3cde5618d07cae39b7841875504471cb2341dc17
Reviewed-on: https://review.typo3.org/50888
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>