[BUGFIX] Fix exception in QuickEdit mode for empty pages
If there is no content on a page the parameter $this->eRParts[1] will be
set to e.g. /new/1 which is not an integer value and it will throw an
exception here on certain dbms, thus let's check that before as there could
not be a history on a new record and skip the query to prevent the error.

Change-Id: If33c5de067b7685e5723ceb7aa1586640abc99d8
Resolves: #78925
Releases: master, 7.6, 6.2
Reviewed-on: https://review.typo3.org/50946
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>