[BUGFIX] Fix exception in QuickEdit mode for empty pages
If there is no content on a page the parameter $this->eRParts[1] will be
set to e.g. /new/1 which is not an integer value and it will throw an
exception here on certain dbms, thus let's check that before as there could
not be a history on a new record and skip the query to prevent the error.

Change-Id: If33c5de067b7685e5723ceb7aa1586640abc99d8
Resolves: #78925
Releases: master, 7.6, 6.2
Reviewed-on: https://review.typo3.org/50910
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Alexander Stehlik <alexander.stehlik@gmail.com>
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Reviewed-by: Michael Oehlhof <typo3@oehlhof.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>