[BUGFIX] Ensure correct field ordering in CSV export
Due to patch https://review.typo3.org/48998/ the whole database record
is fetched from the database instead of only the visible fields. This
screws up the CSV export as this on relies of the field sorting of the
database row. This patch ensures a correct sorting of the row fields
before they are added to the CSV export.

Resolves: #78908
Releases: master, 7.6
Change-Id: I64f0213695c0cd13fecaa96343a5b1d9af1582f4
Reviewed-on: https://review.typo3.org/50944
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Philipp Gampe <philipp.gampe@typo3.org>
Tested-by: Philipp Gampe <philipp.gampe@typo3.org>
Reviewed-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Tested-by: Jigal van Hemert <jigal.van.hemert@typo3.org>