[BUGFIX] Delete relations of tables without delete-field
In an inline relation, when a parent table is not soft delete
aware, and a parent is deleted (hard delete, DB row removed),
attached children need to be hard deleted as well. Otherwise,
those children would be orphaned.

The change updates the delete handling to reflect this and
adds a bunch of functional tests to verify details.

The patch also reveals there are still issues with hard delete
scenarios when more than one inline child is attached.

Resolves: #79014
Releases: main, 13.4
Change-Id: If1686908f2bd056ab4622742eff1a03f9100edbe
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89585
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>