[BUGFIX] Properly handle flexform related exceptions
The FlexFormTools class tries to resolve the datastructure
of flex fields by the given TCA configuration. The flexform
definition can either be set directly in the TCA or it is
fetched from another record.

Example:
TemplaVoilà fetches the data structure from the table
tx_templavoila_datastructure.

When trying to resolve the data structure from another
table and the identifier is invalid (e.g. empty or does
not point to a valid record), several exceptions are
thrown that need to be caught at several points to keep
the user interface accessible.

Other than these mentioned exceptions there are ones
that indicate that the TCA configuration is simply wrong.
These ones are not caught and will still bubble up.

Releases: master, 8.7
Fixes: #79101
Change-Id: I9be921e1425076897a86ebb0b997a998fda7f373
Reviewed-on: https://review.typo3.org/54875
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>