[TASK] Deprecate BackendUtility::getRecordsByField
Deprecate BackendUtility::getRecordsByField() as it has a flawed design
due to passing SQL fragments. This contradicts the goal of using named
parameters for all queries in the core and requires passing the original
QueryBuilder object in addition to the stringified constraint.

Replace all calls to the method with direct usage of the QueryBuilder and
deprecate the method.

Resolves: #79122
Releases: master
Change-Id: I8b040b98e20271aff84ef16fb89b59a406d54003
Reviewed-on: https://review.typo3.org/51078
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Mona Muzaffar <mona.muzaffar@gmx.de>
Tested-by: Mona Muzaffar <mona.muzaffar@gmx.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>