[BUGFIX] Make copy mode of translation modal use localize
...  omitting TransOrigPointer

While translating elements with the localization modal we have to
make sure that the "Copy" mode does more than just a plain copy
followed by an update of the sys_language_uid. Otherwise related
elements will just be copied along but don't get a new language.

This patch replaces the copy action with a new action called
copyToLanguage. It behaves like localize action but does not set
TransOrigPointer fields, so child elements get the correct
language and are not connected to a language parent.

Resolves: #75400
Releases: master, 7.6
Change-Id: If57af3c94c67824ab8bbbc62a8c6bea283e96fca
Reviewed-on: https://review.typo3.org/51109
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>