[TASK] Use $ instead of TYPO3.jQuery
In TYPO3 6.0 we introduced TYPO3.jQuery as a possibility to have jQuery
run side-by-side with prototype.js in a no-conflict mode.

Since TYPO3 7.5, prototype.js was removed, and $ can be used.

PageRenderer now sets TYPO3.jQuery AND $ as shorthand-functionality
and all places in the TYPO3 Core are replaced with the common $
functionality, so TYPO3.jQuery should be avoided.

Resolves: #79221
Releases: master
Change-Id: I9177d7d3ad618fa7bb3a7e2ea483f39d6d4231bc
Reviewed-on: https://review.typo3.org/51212
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>