[BUGFIX] Use POST for storing uc-related Storage in BE
The AJAX call to store uc settings should be handled
as POST request, and not a GET request to ensure that
even long data can be stored on the server and not
run in Firewall/HTTP GET length restrictions.

Additionally a really bad typo did not allow any
POST requests in UserSettingsController as the key
evaluation was wrong.

Resolves: #79266
Releases: master, 7.6
Change-Id: I0fd10090416b1a16f6a9bdc6a07efceccf147336
Reviewed-on: https://review.typo3.org/51263
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>