[BUGFIX] ResourceCompressor does not persist external resources
The condition in the ResourceCompressor for fetching
external files never worked, as it was a simple typo
to update an external resource.

So the first time the local file did not exist (and thus, the md5
did not match) and the file was fetched but due to the
wrong check, the file never got updated.

Resolves: #79269
Releases: master, 7.6
Change-Id: If6942284ccd46f400e23d199e827d4e7a696961d
Reviewed-on: https://review.typo3.org/51274
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>