[!!!][TASK] Remove RTE proc.transformBoldAndItalicTags option
The option transformBoldAndItalicTags is there to convert
<i> and <b> tags from the DB to <strong> and <em> for the RTE.

When saving, all tags are converted back to <i> and <b>
before storing the content in the database again.

This is actually a (weird) shortcut due to historical reasons,
and can easily be set up by the HTMLparser directly using
the tags.[tagname].remap = [newtagname].

This syntax can be used directly to enable this functionality
again

RTE.default.proc {
        # make <strong> and <em> tags when sending to the RTE
    HTMLparser_rte {
            tags {
                b.remap = strong
                i.remap = em
            }
    }
        # make <b> and <i> tags when sending to the DB
    HTMLparser_db {
        tags {
            strong.remap = B
            em.remap = I
        }
    }
}

The option is removed and the explicit (and IMHO more understandable)
syntax is used in the RTE configuration of HTMLArea.

If the InlineElements functionality of HTMLarea is enabled,
then the remapping is not done, as any tag (b, i, strong, em)
is kept as is, becoming more flexible.

In any case, any "unification" if necessary, can also be
done the same way via the HTMLparser for the frontend
when outputting the content.

Resolving: #79300
Releases: master
Change-Id: I354b0ccd1948ea7d018c7135c4694e0c0f9b72b8
Reviewed-on: https://review.typo3.org/51289
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>