[TASK] Include constants in TypoScript cache
In order to work with available constants in TypoScript in Frontend
the parsed TypoScript constants are included in the cache entries
as well.

This ensures a more consistent behaviour of available information
for the TYPO3 Frontend.

This was previously set via $this->backend_info which was
set to "true" in the TYPO3 Backend.

Resolves: #79307
Releases: master
Change-Id: I5f5e024c8df51c4fd43b7bc607855b9b21ebb1dd
Reviewed-on: https://review.typo3.org/51304
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>