[FEATURE] Allow to override PATH_site via environment variables
In order to make it possible to call TYPO3 from outside the document
root (e.g. CLI binaries) the patch allows to set TYPO3_PATH_ROOT via
environment variables.

This way it is possible to call binaries / external applications that
use the TYPO3 system environment builder.

The env variables are already set via the TYPO3 Console composer
plugin, and could also be incorporated in the composer-plugin of
the TYPO3 Core.

Resolves: #79343
Resolves: #79461
Releases: master
Change-Id: I078b050e35ef7de54744318386139859c2160a0b
Reviewed-on: https://review.typo3.org/51331
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Claus Due <claus@phpmind.net>
Tested-by: Claus Due <claus@phpmind.net>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>