[TASK] Deprecate visibility for internal variables in Datahandler class
This patch marks some internal cache variables in the
Datahandler as deprecated since their visibility should be
changed from public to protected or even be replaced
by a runtime cache. Since this can't be done before
TYPO3 v9 they are marked as deprecated in TYPO3 v8
as a preparation.

Fixes: #79441
Releases: master
Change-Id: I7d35792fb81e6f16b3b729a3841c1cfda109952f
Reviewed-on: https://review.typo3.org/51411
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>