[!!!][TASK] EXT:form - Refactor fluid rendering
This patch is a followup of issue #79439.

EXT:form uses "fluid" as the default rendering strategy.
Therefore, EXT:form has to work close with the concepts of fluid to
avoid current and future problems.
Until now, EXT:form tried to reuse a fluid view instance by
reconfiguring the instance on each nesting level, but fluid is not
intended for such a purpose.
Therefore, the templates have to be moved/ changed and some
configuration has to be changed.
The patch breaks EXT:form only for people who are using custom
configurations/ templates.

Resolves: #79464
Releases: master
Change-Id: I6346b888b47a52bcc995c7d4cd3acdc65a1396c8
Reviewed-on: https://review.typo3.org/51442
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Bjoern Jacob <bjoern.jacob@tritum.de>
Tested-by: Andreas Steiger <typo3@andreassteiger.de>
Tested-by: Tobi Kretschmann <tobi@tobishome.de>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>