[BUGFIX] Eliminate gremlins in image cropper
Refactor rendering to use two templates, a layout
and a partial to be more flexible.

Now also force the crop to be applied, even when
the editor does not open the image manipulation.

This is important when only one aspect ratio is
allowed to not allow the editor to save an invalid
state.

Eliminate several quirks in the TypeScript code,
among them:

* reset state when closing the cropper
* reset state when modal is dismissed
* correctly initialize all previews
* show correct size of crop area for big images
* avoid accidental dismiss of the modal

Last but not least add the accidentally removed form engine
wizard html fields again.

Resolves: #79764
Resolves: #79731
Resolves: #79753
Resolves: #79674

Releases: master
Change-Id: I0a24d6418d6263b00c3fbf31901fd7c67e9fc97e
Reviewed-on: https://review.typo3.org/51642
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>