[BUGFIX] Do not use array assignment for string variable
In PHP 7.1 it is not possible anymore to assign a value to
a variable of type string with `[]` appended. In older PHP
versions the variables was silently converted to an array but
in PHP 7.1 a warning is thrown.
Beside that the variable must be of type string anyway and not
an array.

Resolves: #79709
Releases: master
Change-Id: Ie36dcf89cff23e7837d1f559cd05b3f9f448360e
Reviewed-on: https://review.typo3.org/51608
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Johannes Kasberger <johannes.kasberger@reelworx.at>
Tested-by: Johannes Kasberger <johannes.kasberger@reelworx.at>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>