[BUGFIX] Flaky unit tests FrontendUserAuthentication
Unit testing user authentication is still a mess since the class
does way too much to be well testable.
The patch refactors a number of tests that came in with the session
framework test to not rely on mocked $subject anymore, but it also
marks a couple of tests to be skipped for now - those should be
transferred to functional tests which will probably end up to be
much easier to understand and maintainable.

Change-Id: Ie57245abc75f92a5a1a44ab51079f523974e9818
Resolves: #79678
Related: #70316
Releases: master
Reviewed-on: https://review.typo3.org/51620
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>