[TASK] Update require.js version number in bower.json
The version number of require.js should reflect the one of the
already installed version v2.3.2.

Require.js release notes:
2.3.0:
The way dependencies are found for the sugared syntax changed
slightly in #1582. It is now faster and should catch more
legitimate dependency cases. You should not notice any negative
consequences to the change, but given how important that
mechanism is, it was best to rev the minor version to 2.3.
The r.js optimizer also handles larger directory changes better.
If you are using 2.2, you are encouraged to update to 2.3.
alameda was also updated to 1.1.0 to include the same
dependency searching change as requirejs 2.3.

2.3.1:
Only one change from 2.3.0, and it is in the optimizer.
It fixes how the bundled uglifyjs declares the addFile function.
Without the fix, it causes an error in Node 0.10/0.12.
Newer versions of Node are not affected.

2.3.2:
Fixes an issue that prevented the 2.3.x optimizer from running
in Java's 8+'s jjs. There was a change in require.js and a
change in the r.js optimizer.

Resolves: #79738
Releases: master
Change-Id: I117e0809236462fb6888108f1db60d6d1b4254e9
Reviewed-on: https://review.typo3.org/51623
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Josef Glatz <josef.glatz@typo3.org>
Tested-by: Josef Glatz <josef.glatz@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>