[BUGFIX] Remove non existing fields in FormEngine ListOfFieldsContainer
ListOfFieldsContainer of FormEngine render construct is used
to reduce the field list of a full record down to a given list of
fields that should be rendered, only.
It is triggered for instance from the list module if only a single
field of multiple different records should be edited at once.
Until now, the container did not fully validate if a given field
actually exists within the showitem field or palette and basically
delegated the handling down to a different container which in the
end skipped that field if it was not configured in TCA for given
record type.
The patch fixes this by improving the lookup. Field rendering is
no longer delegated down to other containers if the field in
question does not exist within the 'types' section and is not
within a referenced 'palette' of given record type.

Change-Id: I750ebf7c3d87ecb381f7bbe21b63528765841277
Resolves: #79750
Releases: master
Reviewed-on: https://review.typo3.org/51630
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Susanne Moog <susanne.moog@typo3.org>