[TASK] Finalize L10nModeUpdater for enhanced l10n_modes
L10nModeUpdater has to be adjusted to reflect all possibilities that
became available with introducing allowLanguageSynchronization and
the implicit synchronization for the exclude mode.

Instead of cloning the logic again inside the upgrade wizard, the
DataHandler and DataMapResolver is used to determine required changes
as well as performing especially relation resolving and handling.

Resolves: #79768
Releases: master
Change-Id: Id61c419fe9f6aa6fdbbe6f3d6335b0e9c1bf0693
Reviewed-on: https://review.typo3.org/51646
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>