[BUGFIX] Install Tool configuration parser must handle nested config
Since TYPO3_CONF_VARS may contain nested arrays and the array keys
for those subarrays may collide with first-level citizens (e.g. BE)
the current description parsing (parsing of PHP source code file)
fail hard for those cases.
The problem is not solvable without writing a dedicated full PHP parser
in order to have proper notion of nesting levels.
For this reason the regex-based parsing is completely removed and a
new and dedicated file for the option descriptions is introduced,
which stores the description into a PHP array.

This frees us from the buggy PHP code parsing with regex and enables us
in the future to enhance this array with more information.
(e.g. extract data type into a dedicated subarray element)

Resolves: #79920
Releases: master
Change-Id: Idb5d3fdd2dee5f3150b52c66e65d06c028750893
Reviewed-on: https://review.typo3.org/51772
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>