[BUGFIX] Exception when setting "language" to "all" for content elements
Currently you get an error in the backend when you try to set the
language to "all" for a new or existing content element (integer was
expected but null received triggers PHP strict type error).

The reason for this is, that the DataMapProcessor tries to sychronize
elements from tt_content with uid 0.

This patch adds an integration test that content elements without an
existing source do not get synronized.


Resolves: #79853
Releases: master
Change-Id: I67d4b35efe1564c13e9faf085c4848f2d0313998
Reviewed-on: https://review.typo3.org/51804
Tested-by: Claus Due <claus@phpmind.net>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Claus Due <claus@phpmind.net>
Reviewed-by: Johannes Goslar <jogo@kronberger-spiele.de>
Reviewed-by: Morten Pless <mpl@systime.dk>
Reviewed-by: Jasmina Ließmann <code@frauliessmann.de>
Tested-by: Jasmina Ließmann <code@frauliessmann.de>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>