[BUGFIX] DB Check causes PHP warnings and errors
In DatabaseIntegrityCheck the method selectNonEmptyRecordsWithFkeys
also needs to take the BLOB into account or else it will fail with
a PHP warning. Furthermore the returned tableColumns list may contain
quoted fields names if the field name can be interpreted as a keyword
for the used connection platform. This patch ensures the correct
column information is received before further processing is started.

Additionally the "Record Statics" module is fixed by checking for an
empty TCA table field which invokes an exception.

Resolves: #79347
Resolves: #79991
Resolves: #79992
Releases: master
Change-Id: Ie455abd2da90d7dabc773ec345b5c8a87b0b5836
Reviewed-on: https://review.typo3.org/51813
Reviewed-by: Christer V <cvi@systime.dk>
Tested-by: Christer V <cvi@systime.dk>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Claus Due <claus@phpmind.net>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Anders Kostending <aha@systime.dk>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>