[BUGFIX] Fix dynamic variable name parts in Fluid
The dynamic variable syntax does currently work in Fluid standalone,
but not with the variable provider used for TYPO3 CMS (as promised)

To fix this, we copy the necessary code to the CMS variable provider
as it is only a few lines.

Cover that with tests and apply it in the image cropper UI as one example.

Resolves: #79997
Releases: master
Change-Id: I3a811b0d496b12191d2ccbaa9c07525f4cf8f340
Reviewed-on: https://review.typo3.org/51841
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Anders Kostending <aha@systime.dk>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Joerg Boesche <typo3@joergboesche.de>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>