[BUGFIX] Don't update passwords if left untouched
Fixes a bug where editing a backend user record without updating the
password would result in the password being set to
the string literal "*********".

This reverts #79576 because the fix for showing the password hash in the readable
field was wrong and causing this bug.

Instead of forcing the database value in the hidden field to be asterisks,
we now correctly set the type of the human readable field to be password.

This triggers a special handling in the form engine JavaScript, not filling
the human readable field with the database value and switching to type text
when entering a new password.

Change-Id: Ia465293272131c32bbb9fd9b0d3916676e130996
Resolves: #79714
Reverts: #79576
Releases: master, 8.6
Change-Id: I67b91c076e497ae30b96f3ffa4bab89ce33e9501
Reviewed-on: https://review.typo3.org/51887
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>