[BUGFIX] Install tool upgrade wizards fail on postgres
Controller class UpgradeWizard of install tool method performUpdate()
calls getDatabaseConnection() (old TYPO3_DB) and connects. This was
done in the past to receive the list of executed queries during
upgrade and is now implemented in the single upgrade classes directly.

If running postgres, the connect attempt of old database class tries
to connect to a mysql instance and runs into a timeout.

Remove this TYPO3_DB left over from the install tool to successfully
execute upgrade wizards on non-mysql platforms.

Change-Id: Ia87ae4bce1c4f4992527f0fa6fbddebdb6e3733c
Resolves: #80026
Releases: master
Reviewed-on: https://review.typo3.org/51862
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>