[BUGFIX] Add query restrictions for all tables in query
QueryBuilder::getQueriedTables now returns an array with table
alias as a key and table name as value.
If a table doesn't have an alias, both key and value are containing
table name.

This change is needed to make query restriction classes add restrictions
for all tables in case the self inner join is made.
A test covering the case was added to QueryBuilderTest.

Releases: master
Resolves: #79615
Change-Id: Ibf797189334ed86db0fce1a4ef3f0898cb0ada11
Reviewed-on: https://review.typo3.org/51902
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Reviewed-by: Patrick Broens <patrick@patrickbroens.nl>
Tested-by: Patrick Broens <patrick@patrickbroens.nl>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>