[BUGFIX] Felogin: Only wrap header/message if not empty
When adding felogin template the wraps defined by default
of EXT:felogin are always outputted even though it just adds
an empty div/h3 tag around, which looks odd in most installations.

The patch adds a proper required tag to see if the value is actually not
empty.

Resolves: #80110
Releases: master, 7.6
Change-Id: Ia9e4c7d3dc294b30db3f10cdde96f95899aa18ef
Reviewed-on: https://review.typo3.org/51934
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Josef Glatz <josef.glatz@typo3.org>
Tested-by: Josef Glatz <josef.glatz@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>