[BUGFIX] Replace class property caching with runtime cache
PageRepository is now used independantly from TSFE
most of the time, and a new instance is created each time a
content element gets rendered. This means that the current
cache strategy using class properties gets flushed repeatedly
and multiple identical SQL requests to be sent to the DB.

By replacing the class property caching with runtime cache
which survives across instances, the problem is solved.

On some sites this represents a reduction of several hundred
SQL requests.

Change-Id: Ib606ada46c3c262fd9004a12166b79d80bbc5edb
Resolves: #80138
Releases: master
Reviewed-on: https://review.typo3.org/51953
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>