[BUGFIX] Update selected ratio in preview element
When selecting a different ratio and accepting the image manipulation,
not only the image thumbnails, but also the selected ratio should be updated.

Also set the selected ratio correctly when initializing the image manipulation
and set the value even if selected ratio is not defined in TCA but is present
in allowed aspect ratios.

Also always set crop area from saved collection.

Resolves: #80169
Releases: master
Change-Id: I63c24e14b724c3f47ff3101afb4bcb56983c1b28
Reviewed-on: https://review.typo3.org/51979
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Joerg Boesche <typo3@joergboesche.de>
Tested-by: Joerg Boesche <typo3@joergboesche.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>