[BUGFIX] Do not overwrite inlineData in mergeResult
Instead of merging different inlineData results, the
FormResultCompiler has one seemingly leftover line of code
which overwrites the inlineData, leading to the situation
that only the last inlineData is kept.

In the list module when editing multiple fields with IRRE relations,
this leads to the fact that children can only be added
to the last element.

For some reason the same code does not expose the bug to the end user,
most likely due to different handling of the context data.

Nevertheless this obviously wrong line should also be removed in 7.6.

Releases: master, 7.6
Resolves: #80181
Change-Id: I56cef1c56a36d1760a99c239130c6791869a1978
Reviewed-on: https://review.typo3.org/52050
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>