[BUGFIX] Avoid CKEditor RequireJS information loaded in FE
The CKEditor default configuration for RequireJS path
resolving is only relevant in BE, Frontend Page Rendering
needs to do that separately for now, to avoid
loading of too many unnecessary config files if only
requireJS should be used in the Frontend.

The patch changes the functionality to only load
RequireJS in Backend (and not in CLI mode).

Resolves: #80183
Releases: master
Change-Id: I035ddb6bc83ebfb2a8f45de71fadd8f0e0b2d0a3
Reviewed-on: https://review.typo3.org/51987
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Reviewed-by: Christer V <cvi@systime.dk>
Tested-by: Christer V <cvi@systime.dk>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>