[BUGFIX] MM language synchronization is ignored on initial localization
The language synchronization of many-to-many fields (MM) is ignored for
the initial localization command, but is considered again when further
changes happen to source/parent elements of a localized element.

Besides fixing the missing 'config' segment on resolving the foreign
table name of the field to be synchronized, the special handling for
language assignments has been added - albeit it might not been used
in fields different to sys_language_uid.

Change-Id: I8bf02ee117fa41cf5b0425d0e535fced84fb72ca
Resolves: #80211
Releases: master
Reviewed-on: https://review.typo3.org/52004
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Müllenhagen <christianmuellenhagen@yahoo.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>