[BUGFIX] DBAL: Ensure correct mysqli resource for admin_get_tables query
The admin_get_tables method tries to fetch all available database tables
from the default connection. But using the native mode, the query is
executed with the $this->lastHandlerKey link which might not be the
default connection. This patch ensures the correct mysqli resource is
used to resolve the database queries.

Resolves: #80358
Releases: 7.6
Change-Id: I3b8224e5377aad64e556052bbea37dae39ca1bb7
Reviewed-on: https://review.typo3.org/52106
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Elmar Hinz <t3elmar@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>