[TASK] Deprecate TSFE->whichWorkspace parameter
The parameter of the $TSFE->whichWorkspace() method decides
whether the UID of the current workspace should be returned or
the actual "title" of the sys_workspace record.

As this parameter is not in use anymore and clearly makes this function
harder to understand why and when someone could use this, this property
is marked as deprecated.

Instead, fetching the workspace title should happen in the place where it
is needed.

Resolves: #80485
Releases: master
Change-Id: I06512e40c48b35826c0d0a89cc8c4cdcdc8773e6
Reviewed-on: https://review.typo3.org/52184
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>