[TASK] Remove deprecated localizationMode setting for file fields
The current behavior for file fields using TCA type inline is configured
to use localizationMode=select, which is deprecated. Actually files have
been copied along during the localization of the parent record - and
files can then be defined independently for the default language record
and in the translated record. The localizationMode synchronization
behavior was disabled in the user-interface.

With this change the default core behavior will stay the same. If the
localization buttons would have been enabled individually, this now can
be done with the more specific allowLanguageSynchronization feature.

Change-Id: I88668e3d0ce1dd192d32b93f0312796c59d62358
Resolves: #80504
Releases: master
Reviewed-on: https://review.typo3.org/52201
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Thomas Hohn <thomas@hohn.dk>
Tested-by: Thomas Hohn <thomas@hohn.dk>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>