[BUGFIX] Enhance workspace handling of language synchronization
Workspace handling of synchronizing languages between ancestor and
dependent records has been missing. Basically this is all about
resolving the live UID of records that are used for l10n_parent
fields as well as for other relations pointers.

This happens in a very "basic" construct in every versioned
record for editor - even without relations:

When switching to a workspace, one edits an existing page (with
translations) and sets a field with "allowLanguageSynchronization".

FormEngine - in Workspace=1 - utilizes the liveId in data[pages][123]
obviously. When persisting, a new version is created and subsequent
edits use the versioned ID in FormEngine (input name="data[pages][124]")
which are thus not working anymore.

This patch solves the actual problem, making
"allowLanguageSynchronization" work fully now.

The main trick is to always resolve to the live version of the
default langauge (l10n_parent ID) over and over again.

Resolves: #80640
Resolves: #107262
Related: #79244 (previous commit adding tests)
Related: #101285 (previous commit adding tests)
Releases: main, 14.3
Change-Id: I2b44462e6f6f701ccd2f26686ec7400e0023c861
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/52294
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Lisa Kreitz <lisamariakreitz@web.de>
Reviewed-by: Lisa Kreitz <lisamariakreitz@web.de>