[FEATURE] Improved JavaScript Modal API
To improve the usability and flexibility for a unified handling of
overlays in the backend we're opening the existing API for modals to be
more flexible and adjustable to your needs for advanced usage. With the
introduction of the advanded API it is now possible to pass configuration
via a json object. This enables more easy configuation and better
fallbacks if the modal is not correctly configured.

For a unified experience all modals are now centered by default, will be
kept in place automatically and are available in different sizes depending
on your needs. In addition to this there is also a new type for loading
content into an iframe. Buttons now have full support for the TYPO3 Icon
API and data attributes can also be set by configuration.

Resolves: #80579
Releases: master
Change-Id: Iddd260193a55335889db1068fbd8684011db6518
Reviewed-on: https://review.typo3.org/52228
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>