[BUGFIX] FAL: Correctly iterate folder content with offset
Whenever a list of files of a folder is queried with an offset
make sure we handle filtered files correctly by not taking them
into account when searching for the first eligible file for the
result.

A direct seek() to the start position is wrong as those first files
may contain files would be later on filtered.
The solution is to iterate the whole file list and only count
non-filtered elements to reach the correct starting point.

Releases: master, 7.6, 6.2
Resolves: #80589
Change-Id: I79a6d298599d31aa49c36fd51cc5bbe2cf2ccf3a
Reviewed-on: https://review.typo3.org/52287
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>